objId is size_t
git-svn-id: file:///home/roberto/Git/svn@7144 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: abf7d6daeb11eb2c29632bdff0ffab9cf83c56b3

+++ trunk/xmipp/applications/programs/correlation/correlation_main.cpp
@@ -77,7 +77,7 @@ public:
         std::cout << std::endl;
     }
 
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
     {
         Image<double> img;
         img.readApplyGeo(fnImg,mdIn,objId);
