+++ trunk/xmipp/libraries/reconstruction/correct_bfactor.h
@@ -77,7 +77,7 @@ protected:
     void show();
     void defineParams();
     void readParams();
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId);
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId);
 
 public:
